V.1

+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryFilmStorage.java
@@ -1,12 +1,14 @@
 package ru.yandex.practicum.filmorate.storage;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
@@ -45,7 +47,7 @@ public class InMemoryFilmStorage implements FilmStorage {
         validateUserExists(userId);
         validateFilmExists(filmId);
         Film film = films.get(filmId);
-        film.getLikes().remove((Integer) userId);
+        film.getLikes().remove(userId);
     }
 
     @Override
@@ -55,10 +57,7 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     @Override
     public List<Film> getMostLikedFilms(int count) {
-        return films.values().stream()
-                .sorted((f1, f2) -> Integer.compare(f2.getLikes().size(), f1.getLikes().size()))
-                .limit(count)
-                .collect(Collectors.toList());
+        return films.values().stream().sorted((f1, f2) -> Integer.compare(f2.getLikes().size(), f1.getLikes().size())).limit(count).collect(Collectors.toList());
     }
 
     private void validateFilmExists(int filmId) {
